Frontend/meeting-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
openai==1.54.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.0
aiofiles==24.1.0
google-generativeai>=0.8.3
